+++ bindings/nodejs/CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split `Account::send` into `send` and `sendWithParams`;
 - Properties in classes `NewOutputWalletEvent`, `SpentOutputWalletEvent` and `TransactionInclusionWalletEvent` are now private;
 - Utils `computeTokenId` field `tokenSchemeType` renamed to `tokenSchemeKind`;
+- `{Client, SecretManager}::signTransaction` return type from `Payload` to `TransactionPayload`;
 
 ### Removed